api/sam3/pyproject.toml
[project]
name = "sam3"
version = "0.1.0"
description = "SAM3 Video Tracking API for plant CV pipeline"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=0.27.0",
    "litserve>=0.2.8",
    "numpy>=1.24.0",
    "opencv-python>=4.5.0",
    "pillow>=11.2.1",
    "pip>=25.3",
    "torch>=2.7.0",
    "torchvision",
    "transformers>=4.51.2",
    "pytest>=8.0.0",
    "kernels>=0.11.5",
    "av",
]

[tool.uv.sources]
transformers = { git = "https://github.com/huggingface/transformers" }
